Move App component to components directory
